Better callback plugins without exposing apex_plugin internal struct.
