+++ tests/test_custom_plugins.c
@@ -11,7 +11,7 @@
 #include "cmark-gfm.h"
 #include <string.h>
 
-static char * my_plugin_callback(const char *text, __attribute__((unused)) apex_plugin *plugin, apex_plugin_phase_mask phase_mask, __attribute__((unused)) const apex_options *options) {
+static char * my_plugin_callback(const char *text, __attribute__((unused)) const char *id_plugin, apex_plugin_phase_mask phase_mask, __attribute__((unused)) const apex_options *options) {
     if (phase_mask & APEX_PLUGIN_PHASE_PRE_PARSE) {
         const char *suffix = "\n_Hello Sbarex_\n";
         size_t len1 = strlen(text);
@@ -39,57 +39,17 @@ static char * my_plugin_callback(const char *text, __attribute__((unused)) apex_
 }
 
 static void my_plugin_register(apex_plugin_manager *manager, __attribute__((unused)) const apex_options *options) {
-    apex_plugin *plugin1 = init_plugin();
-
-    plugin1->id = strdup("my_plugin");
-    plugin1->title = strdup("my_plugin title");
-    plugin1->author = strdup("sbarex");
-    plugin1->description = NULL;
-    plugin1->homepage = NULL;
-    plugin1->repo = NULL;
-    plugin1->phases = APEX_PLUGIN_PHASE_PRE_PARSE;
-    plugin1->handler_command = NULL;
-    plugin1->priority = 100;
-    plugin1->timeout_ms = 0;
-    plugin1->has_regex = 0;
-    plugin1->replacement = NULL;
-    plugin1->replacement = NULL;
-    plugin1->dir_path = NULL;
-    plugin1->support_dir = NULL;
-    plugin1->pattern = NULL;
-    plugin1->callback = my_plugin_callback;
-
     printf(COLOR_GREEN "✓" COLOR_RESET " Custom plugin register callback called\n");
 
     /* Attach to appropriate phase lists, enforcing per-list id uniqueness */
-    if (plugin_register(manager, plugin1)) {
+    if (apex_plugin_register(manager, "my_plugin", APEX_PLUGIN_PHASE_PRE_PARSE, my_plugin_callback)) {
         printf(COLOR_GREEN "✓" COLOR_RESET " Custom plugin has been registered for pre parse\n");
     } else {
         printf(COLOR_RED "✗" COLOR_RESET " Custom plugin has not been registered for pre parse\n");
     }
 
-    apex_plugin *plugin2 = init_plugin();
-
-    plugin2->id = strdup("my_plugin");
-    plugin2->title = strdup("my_plugin title");
-    plugin2->author = strdup("sbarex");
-    plugin2->description = NULL;
-    plugin2->homepage = NULL;
-    plugin2->repo = NULL;
-    plugin2->phases = APEX_PLUGIN_PHASE_POST_RENDER;
-    plugin2->handler_command = NULL;
-    plugin2->priority = 100;
-    plugin2->timeout_ms = 0;
-    plugin2->has_regex = 0;
-    plugin2->replacement = NULL;
-    plugin2->replacement = NULL;
-    plugin2->dir_path = NULL;
-    plugin2->support_dir = NULL;
-    plugin2->pattern = NULL;
-    plugin2->callback = my_plugin_callback;
-
     /* Attach to appropriate phase lists, enforcing per-list id uniqueness */
-    if (plugin_register(manager, plugin2)) {
+    if (apex_plugin_register(manager, "my_plugin", APEX_PLUGIN_PHASE_POST_RENDER, my_plugin_callback)) {
         printf(COLOR_GREEN "✓" COLOR_RESET " Custom plugin has been registered for post render\n");
     } else {
         printf(COLOR_RED "✗" COLOR_RESET " Custom plugin has not been registered for post render\n");
